working on fixing pull request comments, moved styles, fixed print text, renamed csv/json files, moved utils
