+++ src/generate_csv/verify.ts
@@ -1,10 +1,9 @@
-import { ANSIStyles } from "../print_styles";
-import { isACK, isNAK } from "./utils_csv";
-import { retryJob } from "../utils";
+import { ANSIStyles } from "../constants";
+import { isACK, isNAK, retryJob } from "../utils";
 import {
     stylingVerificationPrompt,
     translationVerificationPrompt,
-} from "./prompts_csv";
+} from "./prompts";
 import type ChatInterface from "../chat_interface/chat_interface";
 import type OverridePrompt from "../interfaces/override_prompt";
