+++ src/translate.ts
@@ -1,5 +1,5 @@
-import { ANSIStyles } from "./print_styles";
 import {
+    ANSIStyles,
     DEFAULT_BATCH_SIZE,
     DEFAULT_REQUEST_TOKENS,
     DEFAULT_TEMPLATED_STRING_PREFIX,
@@ -18,8 +18,8 @@ import PromptMode from "./enums/prompt_mode";
 import RateLimiter from "./rate_limiter";
 import fs from "fs";
 import path, { dirname } from "path";
-import translateCsv from "./generate_csv/generate_csv";
-import translateJson from "./generate_json/generate_json";
+import translateCsv from "./generate_csv/generate";
+import translateJson from "./generate_json/generate";
 import type { TranslationStats, TranslationStatsItem } from "./types";
 import type Chats from "./interfaces/chats";
 import type TranslateDiffOptions from "./interfaces/translate_diff_options";
