+++ src/utils.ts
@@ -1,4 +1,4 @@
-import { ANSIStyles } from "./print_styles";
+import { ANSIStyles } from "./constants";
 import ISO6391 from "iso-639-1";
 import fs from "fs";
 import path from "path";
@@ -114,4 +114,20 @@ export function getTranslationDirectoryKey(
     );
 
     return `${outputPath}:${key}`;
+}
+
+/**
+ * @param response - the message from the LLM
+ * @returns whether the response includes NAK
+ */
+export function isNAK(response: string): boolean {
+    return response.includes("NAK") && !response.includes("ACK");
+}
+
+/**
+ * @param response - the message from the LLM
+ * @returns whether the response only contains ACK and not NAK
+ */
+export function isACK(response: string): boolean {
+    return response.includes("ACK") && !response.includes("NAK");
 }
